requirements.txt
google-generativeai>=0.8.5
fastapi
uvicorn
python-dotenv
